requirements.txt
djangorestframework
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
Pillow
drf-yasg
celery